+++ Issues/11_Miscellaneous/IERC20Indexed/IERC20Indexed.sol
@@ -3,48 +3,65 @@ pragma solidity 0.8.17;
 
 //Event parameters are not indexed.
 interface IERC20 {
-    function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
 
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 
-    function transfer(address recipient, uint amount) external returns (bool);
+    function transfer(address recipient, uint256 amount)
+        external
+        returns (bool);
 
-    function allowance(address owner, address spender) external view returns (uint);
+    function allowance(address owner, address spender)
+        external
+        view
+        returns (uint256);
 
-    function approve(address spender, uint amount) external returns (bool);
+    function approve(address spender, uint256 amount) external returns (bool);
 
     function transferFrom(
         address sender,
         address recipient,
-        uint amount
+        uint256 amount
     ) external returns (bool);
 
-    event Transfer(address from, address to, uint value);
-    event Approval(address owner, address spender, uint value);
+    event Transfer(address from, address to, uint256 value);
+    event Approval(address owner, address spender, uint256 value);
 }
 
 contract ERC20 is IERC20 {
-    uint public override totalSupply;
-    mapping(address => uint) public override balanceOf;
-    mapping(address => mapping(address => uint)) public override allowance;
+    uint256 public override totalSupply;
+    mapping(address => uint256) public override balanceOf;
+    mapping(address => mapping(address => uint256)) public override allowance;
     string public name = "TestIndexedToken";
     string public symbol = "TEST";
     uint8 public decimals = 18;
 
-    function transfer(address recipient, uint amount) external override returns (bool) {
+    function transfer(address recipient, uint256 amount)
+        external
+        override
+        returns (bool)
+    {
         balanceOf[msg.sender] -= amount;
         balanceOf[recipient] += amount;
         emit Transfer(msg.sender, recipient, amount);
         return true;
     }
 
-    function approve(address spender, uint amount) external override returns (bool) {
+    function approve(address spender, uint256 amount)
+        external
+        override
+        returns (bool)
+    {
         allowance[msg.sender][spender] = amount;
         emit Approval(msg.sender, spender, amount);
         return true;
     }
 
-    function transferFrom(address sender,address recipient,uint amount) external override returns (bool) {
+    function transferFrom(
+        address sender,
+        address recipient,
+        uint256 amount
+    ) external override returns (bool) {
         allowance[sender][msg.sender] -= amount;
         balanceOf[sender] -= amount;
         balanceOf[recipient] += amount;
@@ -52,13 +69,13 @@ contract ERC20 is IERC20 {
         return true;
     }
 
-    function mint(uint amount) external {
+    function mint(uint256 amount) external {
         balanceOf[msg.sender] += amount;
         totalSupply += amount;
         emit Transfer(address(0), msg.sender, amount);
     }
 
-    function burn(uint amount) external {
+    function burn(uint256 amount) external {
         balanceOf[msg.sender] -= amount;
         totalSupply -= amount;
         emit Transfer(msg.sender, address(0), amount);
